requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
langchain-groq
python-dotenv
